backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
easyocr
python-multipart
python-dotenv
